fire-cloud/requirements.txt
annotated-types==0.6.0
anyio==4.1.0
awscrt==0.19.1
awsiotsdk==1.19.0
certifi==2023.11.17
deprecation==2.1.0
gotrue==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.26.2
packaging==23.2
postgrest==0.13.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.1.0
supafunc==0.3.1
typing_extensions==4.8.0
websockets==10.4
